+++ src/util/modulegroup/services.go
@@ -16,7 +16,7 @@ func GetModuleGroupsByIDs(moduleGroupIDs []int) ([]models.ModuleGroup, error) {
 }
 
 func CreateModuleGroup(label string, plantID int, locationID int, tds float64, ph float64,
-	humidity float64, lightsOn float64,	lightsOff float64, onAuto bool) error {
+	humidity float64, lightsOn float64, lightsOff float64, onAuto bool) error {
 
 	return repository.CreateModuleGroup(label, plantID, locationID, tds, ph, humidity,
 		lightsOn, lightsOff, onAuto)
@@ -30,8 +30,8 @@ func ToggleAuto(moduleGroupID int) error {
 	return repository.ToggleAuto(moduleGroupID)
 }
 
-func SetEnvironmentParameters(moduleGroupID int, humidity float32, ph float32, tds float32,
-	lightsOnHour float32, lightsOffHour float32) error {
+func SetEnvironmentParameters(moduleGroupID int, humidity float64, ph float64, tds float64,
+	lightsOnHour float64, lightsOffHour float64) error {
 
 	return repository.SetEnvironmentParameters(moduleGroupID, humidity, ph, tds, lightsOnHour,
 		lightsOffHour)
